cilium.io/v2: move files around to reduce imports in API
The API should not have implementation functionality embedded in it. To
reduce the amount of footprint given when importing all of these
libraries we should the functionality logic out of the API package.
