+++ pkg/k8s/apis/cilium.io/v2/client/register_test.go
@@ -14,7 +14,7 @@
 
 // +build !privileged_tests
 
-package v2
+package client
 
 import (
 	"regexp"
